backend/requirements.txt
asgiref==3.11.1
Django==6.0.4
djangorestframework==3.17.1
drf-yasg==1.21.15
gunicorn==25.3.0
packaging==26.1
sqlparse==0.5.5
psycopg[binary]
